Исправила наименования согласно критериям
