+++ src/main.js
@@ -1,13 +1,13 @@
-import { render, RenderPosition, remove} from './render';
-import { MenuItem} from './mock/util';
-import MovieListPresenter from './presenter/movie-list-presenter';
-import FilmStatisticiew from './view/film-stats';
-import FilmsModel from './model/movies-model';
-import FilterModel from './model/filter-model';
-import CommentsModel from './model/comments-model';
-import FilterPresenter from './presenter/filter-presenter';
-import UserStatsView from './view/user-statistic';
-import ApiService from './api-service';
+import { render, RenderPosition, remove } from './utils/render.js';
+import { MenuItem } from './utils/constants.js';
+import MovieListPresenter from './presenter/movie-list-presenter.js';
+import FilmStatisticiew from './view/film-stats.js';
+import FilmsModel from './model/movies-model.js';
+import FilterModel from './model/filter-model.js';
+import CommentsModel from './model/comments-model.js';
+import FilterPresenter from './presenter/filter-presenter.js';
+import UserStatsView from './view/user-statistic.js';
+import ApiService from './service/api-service.js';
 
 const siteMainElement = document.querySelector('.main');
 const containerStatsElement = document.querySelector('.footer__statistics');
@@ -30,7 +30,6 @@ const handleMenuClick = (menuType) => {
       remove(userStatsElement);
       movieListPresenter.destroy();
       movieListPresenter.init();
-      //menuStats.classList.remove('main-navigation__additional--active');
       break;
     case MenuItem.STATISTICS:
       movieListPresenter.destroy();
@@ -40,6 +39,7 @@ const handleMenuClick = (menuType) => {
       break;
   }
 };
+
 const filterPresenter = new FilterPresenter(siteMainElement, filterModel, filmsModel, handleMenuClick);
 
 const renderFooterElement = () => {
