+++ src/utils/constants.js
@@ -1,0 +1,41 @@
+export const isEscKey = (evt) => evt.key === 'Escape' || evt.key === 'Esc';
+
+export const SortType = {
+  DEFAULT: 'default',
+  DATE: 'date',
+  RATING: 'rating',
+};
+
+export const UserAction = {
+  UPDATE_FILM: 'UPDATE_FILM',
+  ADD_COMMENT: 'ADD_COMMENT',
+  DELETE_COMMENT: 'DELETE_COMMENT',
+};
+
+export const UpdateType = {
+  PATCH: 'PATCH',
+  MINOR: 'MINOR',
+  MAJOR: 'MAJOR',
+  INIT: 'INIT',
+};
+
+export const FilterType = {
+  ALL: 'all',
+  FAVORITES: 'favorites',
+  HISTORY: 'history',
+  WATCHLIST: 'watchlist',
+};
+
+export const MenuItem = {
+  FILMS: 'films',
+  STATISTICS: 'statistics',
+};
+
+export const EMOTIONS =[ 'smile', 'sleeping', 'puke', 'angry'];
+
+export const State = {
+  ADDING: 'ADDING',
+  DELETING: 'DELETING',
+  ABORTING: 'ABORTING'
+};
+
