+++ src/view/abstract-view.js
@@ -1,4 +1,4 @@
-import { createElement } from '../render';
+import { createElement } from '../utils/render.js';
 
 const SHAKE_ANIMATION_TIMEOUT = 600;
